Add files via upload
Naive Bayes classifier calculates the probabilities for every factor ( here in case of email example would be Alice and Bob for given input feature). Then it selects the outcome with highest probability.
This classifier assumes the features (in this case we had words as input) are independent. Hence the word naive. Even with this it is powerful algorithm used for
Real time Prediction
Text classification/ Spam Filtering
Recommendation System
So mathematically we can write as,
If we have a certain event E and test actors x1,x2,x3, etc.
We first calculate P(x1| E) , P(x2 | E) … [read as probability of x1 given event E happened] and then select the test actor x with maximum probability value.